add prev fixes and features

+++ src/main/java/com/example/ecm/service/DocumentService.java
@@ -12,6 +12,7 @@ import com.example.ecm.exception.ConflictException;
 import com.example.ecm.mapper.*;
 import com.example.ecm.model.*;
 import com.example.ecm.model.elasticsearch.DocumentElasticsearch;
+import com.example.ecm.model.enums.DocumentState;
 import com.example.ecm.repository.*;
 
 import com.example.ecm.dto.requests.CreateDocumentRequest;
@@ -64,6 +65,9 @@ public class DocumentService {
     private final SignatureMapper signatureMapper;
     private final SearchService searchService;
     private final Base64Manager base64Manager;
+    private final DocumentStateService documentStateService;
+    private final SignatureRequestRepository signatureRequestRepository;
+
 
     private final UserService userService;
 
@@ -138,21 +142,35 @@ public class DocumentService {
         document = document.filter(x -> x.getIsAlive().equals(isAlive));
 
         if (!userPrincipal.isAdmin()) {
-            document = document.filter(d -> d.getUser().getId().equals(userPrincipal.getId()));
+            document = document
+                    .filter(d ->
+                            Objects.equals(d.getUser().getId(), userPrincipal.getId())
+                                    ||
+                                    d.getDocumentVersions().stream()
+                                            .anyMatch(version ->
+                                                    signatureRequestRepository.existsByUserToIdAndDocumentVersionId(
+                                                            userPrincipal.getId(), version.getId()
+                                                    )
+                                            ));
         }
 
         Document doc = document.orElseThrow(() -> new NotFoundException("Document with id: " + id + " not found"));
 
         CreateDocumentResponse response = documentMapper.toCreateDocumentResponse(doc);
 
-        return getCreateDocumentResponse(doc, response);
+        return getCreateDocumentResponse(doc, response, userPrincipal);
     }
 
-    public CreateDocumentVersionResponse getDocumentVersionById(Long documentId, Long versionId, Boolean isAlive) {
+    public CreateDocumentVersionResponse getDocumentVersionById(Long documentId, Long versionId, Boolean isAlive, UserPrincipal userPrincipal) {
         Optional<DocumentVersion> documentVersion = documentVersionRepository.findByDocumentIdAndVersionId(documentId, versionId);
 
-        if (isAlive) {
-            documentVersion = documentVersion.filter(DocumentVersion::getIsAlive);
+        documentVersion = documentVersion.filter(v -> v.getDocument().equals(isAlive));
+
+        if (!userPrincipal.isAdmin()) {
+            documentVersion = documentVersion.filter(version ->
+                    Objects.equals(version.getDocument().getUser().getId(), userPrincipal.getId())
+                            ||
+                            signatureRequestRepository.existsByUserToIdAndDocumentVersionId(userPrincipal.getId(), version.getId()));
         }
 
         DocumentVersion version = documentVersion.orElseThrow(() -> new NotFoundException("Document Version with id: " + versionId + " or Document id " + documentId + " not found"));
@@ -182,39 +200,99 @@ public class DocumentService {
                 .distinct()
                 .collect(Collectors.toList());
 
-        int start = page * size;
-        int end = Math.min(start + size, latestVersions.size());
-
-        Stream<Document> documentStream = documentRepository.findAllById(documentIds.subList(start, end)).stream();
+        Stream<Document> documentStream = documentRepository.findAllById(documentIds).stream();
 
         documentStream = documentStream.filter(x -> x.getIsAlive().equals(isAlive));
 
         if (!userPrincipal.isAdmin()) {
-            documentStream = documentStream.filter(d -> d.getUser().getId().equals(userPrincipal.getId()));
+            documentStream = documentStream
+                    .filter(document ->
+                            Objects.equals(document.getUser().getId(), userPrincipal.getId())
+                                    ||
+                                    document.getDocumentVersions().stream()
+                                            .anyMatch(version ->
+                                                    signatureRequestRepository.existsByUserToIdAndDocumentVersionId(
+                                                            userPrincipal.getId(), version.getId()
+                                                    )
+                                            ));
         }
 
-        List<CreateDocumentResponse> createDocumentResponses = documentStream
+
+        List<Document> filteredDocuments = documentStream.toList();
+
+
+        int start = page * size;
+        int end = Math.min(start + size, filteredDocuments.size());
+        List<Document> paginatedDocuments = filteredDocuments.subList(start, end);
+
+
+        List<CreateDocumentResponse> createDocumentResponses = paginatedDocuments.stream()
                 .map(document -> {
                     CreateDocumentResponse response = documentMapper.toCreateDocumentResponse(document);
-                    return getCreateDocumentResponse(document, response);
+                    return getCreateDocumentResponse(document, response, userPrincipal);
                 })
                 .toList();
+
         return new PageImpl<>(
                 createDocumentResponses, PageRequest.of(page, size),
-                latestVersions.size()
+                filteredDocuments.size()
         ).getContent();
     }
 
-    private CreateDocumentResponse getCreateDocumentResponse(Document document, CreateDocumentResponse response) {
-        response.setDocumentVersions(document.getDocumentVersions().stream()
+    public int getCountDocuments(Boolean showOnlyAlive, UserPrincipal userPrincipal) {
+
+        List<DocumentVersion> latestVersions = documentVersionRepository.findLatestDocumentVersions();
+
+        latestVersions.sort(Comparator.comparing(DocumentVersion::getCreatedAt)
+                .reversed());
+
+        List<Long> documentIds = latestVersions.stream()
+                .map(version -> version.getDocument().getId())
+                .distinct()
+                .collect(Collectors.toList());
+
+
+        Stream<Document> documentStream = documentRepository.findAllById(documentIds.subList(0, latestVersions.size())).stream();
+
+        if (showOnlyAlive) {
+            documentStream = documentStream.filter(Document::getIsAlive);
+        } else {
+            documentStream = documentStream.filter(document -> document.getIsAlive().equals(Boolean.FALSE));
+        }
+
+        if (!userPrincipal.isAdmin()) {
+            documentStream = documentStream
+                    .filter(document ->
+                            Objects.equals(document.getUser().getId(), userPrincipal.getId())
+                                    ||
+                                    document.getDocumentVersions().stream()
+                                            .anyMatch(version ->
+                                                    signatureRequestRepository.existsByUserToIdAndDocumentVersionId(
+                                                            userPrincipal.getId(), version.getId()
+                                                    )
+                                            ));
+        }
+
+        return documentStream.toList().size();
+
+    }
+
+
+    private CreateDocumentResponse getCreateDocumentResponse(Document document, CreateDocumentResponse response, UserPrincipal userPrincipal) {
+        Stream<DocumentVersion> documentStream = document.getDocumentVersions().stream();
+        if (!userPrincipal.isAdmin()) {
+            documentStream = documentStream.filter(version ->
+                    Objects.equals(version.getDocument().getUser().getId(), userPrincipal.getId())
+                            ||
+                            signatureRequestRepository.existsByUserToIdAndDocumentVersionId(userPrincipal.getId(), version.getVersionId()));
+        }
+        response.setDocumentVersions(documentStream
                 .map(version -> {
                     CreateDocumentVersionResponse versionResponse = documentVersionMapper.toCreateDocumentVersionResponse(version);
                     String base64Content = minioService.getBase64DocumentByName(version.getId() + "_" + version.getTitle());
                     versionResponse.setBase64Content(base64Content);
                     return versionResponse;
-                })
-                .toList());
-
+                }).toList());
         return response;
     }
 
@@ -230,6 +308,10 @@ public class DocumentService {
         Document document = documentRepository.findById(id)
                 .filter(Document::getIsAlive)
                 .orElseThrow(() -> new NotFoundException("Document with id: " + id + " not found"));
+        if (!documentStateService.checkTransition(document, DocumentState.DELETED)) {
+            throw new ConflictException("You cannot delete document with id: " + id + " check available transitions");
+        }
+        document.setState(DocumentState.DELETED);
         document.setIsAlive(false);
         documentRepository.save(document);
         searchService.deleteByDocumentVersionId(document.getDocumentVersions().getLast().getId());
@@ -239,6 +321,10 @@ public class DocumentService {
         Document document = documentRepository.findById(id)
                 .filter(d -> !d.getIsAlive())
                 .orElseThrow(() -> new NotFoundException("Deleted Document with id: " + id + " not found"));
+        if (!documentStateService.checkTransition(document, DocumentState.CREATED)) {
+            throw new ConflictException("You cannot recover document with id: " + id + " check available transitions");
+        }
+        document.setState(DocumentState.CREATED);
         document.setIsAlive(true);
         documentRepository.save(document);
         searchService.recoverByDocumentVersionId(document.getDocumentVersions().getLast().getId());
@@ -262,6 +348,11 @@ public class DocumentService {
                 .filter(Document::getIsAlive)
                 .orElseThrow(() -> new NotFoundException("Document with id: " + id + " not found"));
 
+        if (!documentStateService.checkTransition(document, DocumentState.MODIFIED)) {
+            throw new ConflictException("You cannot modify document with id: " + id + " check available transitions");
+        }
+
+        document.setState(DocumentState.MODIFIED);
         DocumentVersion documentVersion = documentVersionMapper.toDocumentVersion(createDocumentVersionRequest);
         documentVersion.setVersionId((long) (document.getDocumentVersions().size() + 1));
         documentVersion.setCreatedAt(LocalDateTime.now());
@@ -330,28 +421,64 @@ public class DocumentService {
      * @return объект {@link CreateDocumentVersionResponse}, содержащий обновленные данные о версии документа
      * @throws NotFoundException если версия документа с указанным ID не найдена
      */
+    @Transactional
     public CreateDocumentVersionResponse patchDocument(Long id, PatchDocumentVersionRequest request) {
-        Document document= documentRepository.findById(id).orElseThrow(() -> new NotFoundException("Document with id: " + id + " not found"));
+        Document document = documentRepository.findById(id).orElseThrow(() -> new NotFoundException("Document with id: " + id + " not found"));
+        if (!documentStateService.checkTransition(document, DocumentState.MODIFIED)) {
+            throw new ConflictException("You cannot modify document with id: " + id + " check available transitions");
+        }
 
+        document.setState(DocumentState.MODIFIED);
         DocumentVersion documentVersion = document.getDocumentVersions().getLast();
 
+
+        DocumentVersion newVersion2 = new DocumentVersion();
+        newVersion2.setDocument(document);
+        newVersion2.setVersionId(documentVersion.getVersionId() + 1);
+
+        if (request.getTitle() != null) {
+            newVersion2.setTitle(request.getTitle());
+        } else {
+            newVersion2.setTitle(documentVersion.getTitle());
+        }
+
+        setValues(documentVersion.getValues().entrySet().stream()
+                .map(entry -> {
+                    SetValueRequest setValueRequest = new SetValueRequest();
+                    setValueRequest.setAttributeName(entry.getKey().getName());
+                    setValueRequest.setValue(entry.getValue().getValue());
+                    return setValueRequest;
+                }).toList(), newVersion2);
+
+        newVersion2.setDescription(documentVersion.getDescription());
+        newVersion2.setCreatedAt(LocalDateTime.now());
+        newVersion2.setIsAlive(true);
+
+
+        DocumentVersion newVersion = documentVersionRepository.save(newVersion2);
+        newVersion = documentVersionRepository.findById(newVersion.getId()).orElse(null);
+
+
         if (request.getDescription() != null) {
-            documentVersion.setDescription(request.getDescription());
+            newVersion.setDescription(request.getDescription());
         }
         if (request.getTitle() != null) {
-            CreateDocumentVersionRequest requestDocumentVersion = documentVersionMapper.toCreateDocumentVersionRequest(documentVersion, minioService.getBase64DocumentByName(documentVersion.getId() + "_" + documentVersion.getTitle()));
-            documentVersion.setTitle(request.getTitle());
-            requestDocumentVersion.setTitle(documentVersion.getTitle());
-            minioService.addDocument(documentVersion.getId(), requestDocumentVersion);
+            CreateDocumentVersionRequest requestDocumentVersion = documentVersionMapper.toCreateDocumentVersionRequest(newVersion, minioService.getBase64DocumentByName(documentVersion.getId() + "_" + documentVersion.getTitle()));
+            newVersion.setTitle(request.getTitle());
+            requestDocumentVersion.setTitle(newVersion.getTitle());
+            minioService.addDocument(newVersion.getId(), requestDocumentVersion);
         }
+
         if (request.getBase64Content() != null) {
-            minioService.addDocument(documentVersion.getId(), documentVersionMapper.toCreateDocumentVersionRequest(documentVersion, request.getBase64Content()));
+            minioService.addDocument(newVersion.getId(), documentVersionMapper.toCreateDocumentVersionRequest(newVersion, request.getBase64Content()));
+        } else {
+            minioService.addDocument(newVersion.getId(), documentVersionMapper.toCreateDocumentVersionRequest(newVersion, minioService.getBase64DocumentByName(documentVersion.getId() + "_" + newVersion.getTitle())));
         }
         if (request.getValues() != null) {
-            setValues(request.getValues(), documentVersion);
+            setValues(request.getValues(), newVersion);
         }
-        CreateDocumentVersionResponse response = documentVersionMapper.toCreateDocumentVersionResponse(documentVersionRepository.save(documentVersion));
-        response.setBase64Content(minioService.getBase64DocumentByName(documentVersion.getId() + "_" + documentVersion.getTitle()));
+        CreateDocumentVersionResponse response = documentVersionMapper.toCreateDocumentVersionResponse(documentVersionRepository.save(newVersion));
+        response.setBase64Content(minioService.getBase64DocumentByName(newVersion.getId() + "_" + newVersion.getTitle()));
 
         documentVersionRepository.save(documentVersion);
         return response;
@@ -372,4 +499,5 @@ public class DocumentService {
 
         return commentMapper.toAddCommentResponse(comment);
     }
+
 }
